+++ src/lib.rs
@@ -260,8 +260,8 @@ pub use bitreader::{BitReader, BitReaderError};
 pub use cabac::{init_type, CabacEngine, CabacError, ContextModel};
 pub use ctx_init::SliceContexts;
 pub use deblock::{
-    derive_boundary_strength, derive_edge_flags, BoundaryStrength, EdgeFlags, EdgeType,
-    TransformSplit,
+    derive_boundary_strength, derive_edge_flags, filter_cu_edges, BoundaryStrength, DeblockCu,
+    DeblockCuParams, EdgeFlags, EdgeType, TransformSplit,
 };
 pub use hrd::{
     CpbEntry, HrdCommonInfo, HrdError, HrdParameters, SubLayerHrd, SubLayerHrdParameters,